requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
bcrypt==4.1.1
bleach==6.1.0
itsdangerous==2.2.0

# OpenSearch
opensearch-py==2.4.2
opensearch-dsl==2.1.0

# Redis
redis==5.0.1
hiredis==2.2.3

# RabbitMQ (optional)
# aio-pika==9.3.1
# pika==1.3.2

# Korean Language Processing (optional)
# konlpy==0.6.0
# kiwipiepy==0.17.1
# sentence-transformers==2.2.2

# Translation (optional - has conflicts)
# googletrans==4.0.0-rc1  # Conflicts with httpx version
# deep-translator==1.11.4

# Data Processing
pandas==2.1.3
numpy==1.24.3
python-dateutil==2.8.2

# Utilities
httpx>=0.13.3  # Flexible version to avoid conflicts
orjson==3.9.10
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock==3.12.0
faker==20.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14